Discrete

Discrete

BeliefUpdaters contains a default implementation of a discrete bayesian filter. The updater is defined with the DiscreteUpdater type. The DiscreteBelief type is provided to represent discrete beliefs for discrete states POMDPs.

A convenience function uniform_belief is probided to create a DiscreteBelief with equal probability for each state.

BeliefUpdaters.DiscreteBelief — Type.

DiscreteBelief

A belief specified by a probability vector.

Normalization of b is NOT enforced at all times, but the DiscreteBeleif(pomdp, b) constructor will warn, and update(...) always returns a belief with normalized b.

Constructor:

DiscreteBelief(pomdp, b::Vector{Float64}; check::Bool=true)

Fields

- `pomdp` : the POMDP problem  
- `state_list` : a vector of ordered states
- `b` : the probability vector

source

BeliefUpdaters.DiscreteUpdater — Type.

DiscreteUpdater

An updater type to update discrete belief using the discrete Bayesian filter.

Fields

- pomdp <: POMDP

source

BeliefUpdaters.uniform_belief — Method.

 uniform_belief(pomdp)

Return a DiscreteBelief with equal probability for each state.

source